refactor: reorganize chart-related utilities into a dedicated Utils directory
